GROUP-105 Updated OpenAiApiConfig jitterFactor to be of appropriate type double

+++ src/main/java/org/grouphq/groupservice/config/OpenAiApiConfig.java
@@ -42,6 +42,6 @@ public class OpenAiApiConfig {
         private int maxAttempts;
         private int initialDelay;
         private int maxDelay;
-        private int jitterFactor;
+        private double jitterFactor;
     }
 }
